Sort projects alphabetically by name within each section
