+++ config.rb
@@ -1,20 +1,4 @@
-module ToolboxHelpers
-  def categories
-    @categories ||= begin
-      File.open("data.json", "r") do |f|
-        JSON.parse(f.read)["categories"]
-      end
-    end
-  end
-
-  def tags
-    @tags ||= categories.map {|c|
-      c['wrappers'].map {|w|
-        w['tags']
-      }
-    }.flatten.uniq.sort_by(&:downcase)
-  end
-end
+require 'lib/toolbox_helpers'
 
 helpers ToolboxHelpers
